+++ crates/zng-app/src/event/events.rs
@@ -1,5 +1,7 @@
+use hashbrown::HashSet;
 use zng_app_context::app_local;
 use zng_time::INSTANT_APP;
+use zng_txt::Txt;
 
 use crate::update::{UpdatesTrace, UPDATES};
 
@@ -11,16 +13,21 @@ app_local! {
 
 pub(crate) struct EventsService {
     updates: Mutex<Vec<EventUpdate>>, // not locked, used to make service Sync.
-    commands: Vec<Command>,
+    commands: CommandSet,
     register_commands: Vec<Command>,
+    l10n: EventsL10n,
 }
-
+enum EventsL10n {
+    Pending(Vec<(&'static str, Command, &'static str, CommandMetaVar<Txt>)>),
+    Init(Box<dyn Fn(&'static str, Command, &'static str, CommandMetaVar<Txt>) + Send + Sync>),
+}
 impl EventsService {
     const fn new() -> Self {
         Self {
             updates: Mutex::new(vec![]),
-            commands: vec![],
+            commands: HashSet::with_hasher(BuildFxHasher),
             register_commands: vec![],
+            l10n: EventsL10n::Pending(vec![]),
         }
     }
 
@@ -46,16 +53,31 @@ impl EventsService {
     }
 }
 
+/// Const rustc-hash hasher.
+#[derive(Clone, Default)]
+pub struct BuildFxHasher;
+impl std::hash::BuildHasher for BuildFxHasher {
+    type Hasher = rustc_hash::FxHasher;
+
+    fn build_hasher(&self) -> Self::Hasher {
+        rustc_hash::FxHasher::default()
+    }
+}
+
+/// Registered commands set.
+pub type CommandSet = HashSet<Command, BuildFxHasher>;
+
 /// App events and commands service.
 pub struct EVENTS;
 impl EVENTS {
     /// Commands that had handles generated in this app.
     ///
     /// When [`Command::subscribe`] is called for the first time in an app, the command gets added
-    /// to this list after the current update.
+    /// to this list after the current update, if the command is app scoped it remains on the list for
+    /// the lifetime of the app, if it is window or widget scoped it only remains while there are handles.
     ///
     /// [`Command::subscribe`]: crate::event::Command::subscribe
-    pub fn commands(&self) -> Vec<Command> {
+    pub fn commands(&self) -> CommandSet {
         EVENTS_SV.read().commands.clone()
     }
 
@@ -76,12 +98,8 @@ impl EVENTS {
         {
             let ev = &mut *ev;
             for cmd in ev.register_commands.drain(..) {
-                if cmd.update_state() {
-                    if ev.commands.iter().any(|c| c == &cmd) {
-                        tracing::error!("command `{cmd:?}` is already registered")
-                    } else {
-                        ev.commands.push(cmd);
-                    }
+                if cmd.update_state() && !ev.commands.insert(cmd) {
+                    tracing::error!("command `{cmd:?}` is already registered")
                 }
             }
         }
@@ -98,5 +116,47 @@ impl EVENTS {
             }
         }
         updates
+    }
+}
+
+/// EVENTS L10N integration.
+#[allow(non_camel_case_types)]
+pub struct EVENTS_L10N;
+impl EVENTS_L10N {
+    pub(crate) fn init_meta_l10n(&self, file: &'static str, cmd: Command, meta_name: &'static str, txt: CommandMetaVar<Txt>) {
+        {
+            let sv = EVENTS_SV.read();
+            if let EventsL10n::Init(f) = &sv.l10n {
+                f(file, cmd, meta_name, txt);
+                return;
+            }
+        }
+
+        let mut sv = EVENTS_SV.write();
+        match &mut sv.l10n {
+            EventsL10n::Pending(a) => a.push((file, cmd, meta_name, txt)),
+            EventsL10n::Init(f) => f(file, cmd, meta_name, txt),
+        }
+    }
+
+    /// Register a closure that is called to localize command metadata.
+    ///
+    /// The closure arguments are:
+    ///
+    /// * `file` is the command declaration `@l10n: "file"` value or is empty if `@l10n` was set to something else.
+    /// * `cmd` is the command, the command event name should be used as key.
+    /// * `meta` is the metadata name, for example `"name"`, should be used as attribute.
+    /// * `txt` is text variable that must be set with the translation.
+    pub fn init_l10n(&self, localize: impl Fn(&'static str, Command, &'static str, CommandMetaVar<Txt>) + Send + Sync + 'static) {
+        let mut sv = EVENTS_SV.write();
+        match &mut sv.l10n {
+            EventsL10n::Pending(a) => {
+                for (f, k, a, t) in a.drain(..) {
+                    localize(f, k, a, t);
+                }
+            }
+            EventsL10n::Init(_) => panic!("EVENTS_L10N already has a localizer"),
+        }
+        sv.l10n = EventsL10n::Init(Box::new(localize));
     }
 }